backend chicken

+++ backend/src/main/java/de/neuefischer/backend/dto/ChickenDto.java
@@ -3,8 +3,8 @@ package de.neuefischer.backend.dto;
 import java.time.LocalDate;
 
 public record ChickenDto(
-        String race, Integer weightInFirstDay, Integer expectedSlaughterWeight,
-        Integer expectedSlaughterAge,Double feedConversion, String hatchery,
+        String race, Double weightInFirstDay, Double expectedSlaughterWeight,
+        Integer expectedSlaughterAge, Double feedConversion, String hatchery,
         String hatchDay
 ) {
 }
